finance/tracking/pyproject.toml
[project]
name = "tracking"
dependencies = [
    "Flask",
    "Flask-Migrate",
    "SQLAlchemy",
    "psycopg2-binary",
    "python-dotenv"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
